% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/machines.R
\docType{data}
\name{machines}
\alias{machines}
\title{Machines}
\format{
A data frame with 300 rows and 3 variables:
\describe{
\item{name}{Titles A, B, or C (type of machine)}
\item{value}{Elapsed time to follow-up}
\item{failure}{Indicator of machine failure}
}
}
\source{
Generated for package.
}
\usage{
machines
}
\description{
This dataset contains some example data for personYears.
}
\keyword{datasets}
