% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personYears.R
\name{personYears}
\alias{personYears}
\title{personYears}
\usage{
personYears(
  time1,
  time2 = NULL,
  event = NULL,
  group = NULL,
  unit = "year",
  timeCut = NULL,
  rate = F
)
}
\arguments{
\item{time1}{(numeric vector) Times to follow-up by default. If time2 is stated, time1 is the start time.}

\item{time2}{(numeric vector) NULL by default. If stated, time2 is the end time.}

\item{event}{(numeric vector) NULL by default. If stated, event indicates an event occured or did not occured.}

\item{group}{(factor vector) NULL by default. If stated, group stratifies the dataset by length(unique(group)) stratas.}

\item{unit}{(string vector 1L) "year" by default. State the unit of time1, time2 with either c("year", "month", "week", "day"). This scales time1,time2 to years.}

\item{timeCut}{(numeric vector) NULL by default. If desired, cuts elapsed time into intervals. Can be unique breakpoints or number of intervals. Creates a second table.}

\item{rate}{(logical vector 1L) F by default. If = T, adds a new variable column displaying the rate n events per 1000 person years.}
}
\value{
A list of one table by default. If !is.null(group), a list of two tables.
}
\description{
Create a simple person-years table.
}
